Add authentication&authorization logic to Board, User modules. and close #4, close #6
Fix Dto username length, omit PK from updateDTO

업데이트용 DTO에 PK가 들어가있어 추후 삽입시 문제가 될 수 있는 부분을 수정하였음

Make UserService to check where user to UD is same

User entity 수정, 삭제하려는 사용자가 entity와 같은 사용자인지 확인하도록 수정하였음

Fix Board module to Only authorized user can modify entity

글을 작성한 사용자만 글을 변경할 수 있도록 제약조건을 추가하였음. 이를 위해 JwtAuthGuard도 추가함

+++ src/board/board.service.ts
@@ -1,10 +1,11 @@
-import { Injectable, InternalServerErrorException } from '@nestjs/common';
+import { Injectable, InternalServerErrorException, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Connection, DeleteResult, Repository, UpdateResult } from 'typeorm';
 import { Board } from './entities/board.entity';
 import { CreateBoardDto } from './dto/create-board.dto';
 import { UpdateBoardDto } from './dto/update-board.dto';
 import { User } from 'src/user/entities/user.entity';
+import { use } from 'passport';
 
 @Injectable()
 export class BoardService {
@@ -27,24 +28,34 @@ export class BoardService {
     }
 
     async insertOne(username: string, createBoardDto: CreateBoardDto): Promise<Board> {
-
         const newBoard = this.boardRepository.create({ username: username, ...createBoardDto });
         console.log(newBoard);
 
         return await this.boardRepository.save(newBoard);
     }
 
-    async updateOne(id: number, updateBoardDto: UpdateBoardDto): Promise<Board> {
+    async updateOne(username: string, id: number, updateBoardDto: UpdateBoardDto): Promise<Board> { 
+        await this.checkAuthority(username, id);
+
         await this.boardRepository.update(id, updateBoardDto);
         return this.boardRepository.findOne(id);
     }
 
-    async deleteOne(id: number): Promise<{ deleted: boolean, message?: string }> {
+    async deleteOne(username: string, id: number): Promise<{ deleted: boolean, message?: string }> {
+        await this.checkAuthority(username, id);
+        
         try {
             await this.boardRepository.delete(id);
             return {deleted: true};
         } catch(e) {
             return {deleted: false, message: e.message}
+        }
+    }
+
+    private async checkAuthority(username: string, boardId: number): Promise<void> {
+        const currentBoard = await this.boardRepository.findOne(boardId);
+        if(currentBoard?.username !== username) {
+            throw new UnauthorizedException();
         }
     }
 }
